% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_shinyapps_file.R
\name{add_shinyapps_file}
\alias{add_shinyapps_file}
\title{Creates an app.R file that can be deployed to shinyapps.io
Inspired by golem::add_rstudio_files()}
\usage{
add_shinyapps_file(
  pkg = pkgload::pkg_path(),
  open = rlang::is_interactive(),
  runapp_funct = "run_app"
)
}
\arguments{
\item{pkg}{pacakge directory}

\item{open}{logical. Should file be opened afterwards}
}
\description{
Creates an app.R file that can be deployed to shinyapps.io
Inspired by golem::add_rstudio_files()
}
