% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_path.R
\name{format_gh_url}
\alias{format_gh_url}
\title{simply format github URL to remove .git extension and https or ssh header}
\usage{
format_gh_url(url)
}
\arguments{
\item{url}{url to reformat}
}
\value{
a character vector
}
\description{
simply format github URL to remove .git extension and https or ssh header
}
\examples{
format_gh_url("https://github.com/Name/repo.git")
}
