% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_path.R
\name{github_path}
\alias{github_path}
\title{Get Github path from repository}
\usage{
github_path(pkg_path = pkgload::pkg_path())
}
\arguments{
\item{pkg_path}{path to repository}
}
\value{
a character vector with Github repository information
}
\description{
Get Github path from repository
}
\examples{
github_path()
}
