% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_exported}
\alias{is_exported}
\title{Is fct exported from package stored in pkg_path}
\usage{
is_exported(fct, pkg_path = pkgload::pkg_path())
}
\arguments{
\item{fct}{a function name (run_app for example)}

\item{pkg_path}{where the package is stored}
}
\value{
a boolean
}
\description{
Is fct exported from package stored in pkg_path
}
\examples{
is_exported("run_app")
}
