% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy_compiled_app.R
\name{deploy_compiled_app}
\alias{deploy_compiled_app}
\title{Deploy a compiled app to Shinyapps.io}
\usage{
deploy_compiled_app(pkg = pkgload::pkg_path(), new_deploy = FALSE)
}
\arguments{
\item{pkg}{path to package to upload}
}
\description{
Deploy a compiled app to Shinyapps.io
}
